--- a/.github/workflows/examples.yaml
+++ b/.github/workflows/examples.yaml
@@ -17,7 +17,7 @@ on:
 
 env:
   EMSDK_VERSION: "3.1.40"
-  TINYGO_VERSION: "0.27.0"
+  TINYGO_VERSION: "0.28.1"
   ZIG_VERSION: "0.11.0-dev.3334+cd1417dbd"
 
 concurrency: