… examples/allocation/tinygo/testdata/greet.wasm
… examples/basic/testdata/add.wasm
… examples/cli/testdata/cli.wasm
… imports/wasi_snapshot_preview1/example/testdata/tinygo/cat.wasm
… internal/integration_test/bench/testdata/case.wasm