This is slower to compile, which I suspect is a downside to supporting more reflection features. Similarly, except trivial functions, the wasm is larger than before.

```
$ go test -run='^$' -bench '^BenchmarkAllocation.*' ./internal/integration_test/vs/compiler -count=6 > v0.27.0.txt
$ go test -run='^$' -bench '^BenchmarkAllocation.*' ./internal/integration_test/vs/compiler -count=6 > v0.28.1.txt
$ benchstat  v0.27.0.txt v0.28.1.txt
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/vs/compiler
                          │ v0.27.0.txt  │            v0.28.1.txt             │
                          │    sec/op    │   sec/op     vs base               │
Allocation/Compile-12       3.124m ±  1%   4.227m ± 2%  +35.30% (p=0.002 n=6)
Allocation/Instantiate-12   126.4µ ± 10%   119.1µ ± 4%        ~ (p=0.132 n=6)
Allocation/Call-12          1.260µ ±  0%   1.420µ ± 1%  +12.66% (p=0.002 n=6)
geomean                     79.23µ         89.40µ       +12.84%

                          │ v0.27.0.txt  │              v0.28.1.txt              │
                          │     B/op     │     B/op      vs base                 │
Allocation/Compile-12       1.292Mi ± 0%   1.640Mi ± 0%  +26.90% (p=0.002 n=6)
Allocation/Instantiate-12   230.5Ki ± 0%   231.2Ki ± 0%   +0.30% (p=0.002 n=6)
Allocation/Call-12            48.00 ± 0%     48.00 ± 0%        ~ (p=1.000 n=6) ¹
geomean                     24.27Ki        26.30Ki        +8.37%
¹ all samples are equal

                          │ v0.27.0.txt │             v0.28.1.txt              │
                          │  allocs/op  │  allocs/op   vs base                 │
Allocation/Compile-12       1.595k ± 0%   1.775k ± 0%  +11.29% (p=0.002 n=6)
Allocation/Instantiate-12    508.0 ± 0%    508.0 ± 0%        ~ (p=1.000 n=6) ¹
Allocation/Call-12           5.000 ± 0%    5.000 ± 0%        ~ (p=1.000 n=6) ¹
geomean                      159.4         165.2        +3.63%
¹ all samples are equal
```